# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manypost API",
    "version": "0.0.1",
    "description": "API do manypost — agendamento/publicação multicanal. Humanos usam sessão Clerk; máquinas usam `Authorization: Bearer mp_live_…`. Erros seguem problem+json (RFC 9457): o campo `title` carrega o código estável do erro."
  },
  "servers": [
    {
      "url": "/",
      "description": "mesma origem (proxy do Next)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "sessão Clerk humana — Authorization: Bearer <token>"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key mp_live_ para superfícies de máquina — Authorization: Bearer <chave>"
      }
    },
    "schemas": {
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "lastUsedAt",
          "revokedAt",
          "createdAt"
        ]
      },
...
      "PublicationFeed": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedItem"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "cursor keyset da próxima página"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/api-keys": {
      "get": {
        "tags": [
          "api-keys"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Lista as API keys da organização",
        "responses": {
          "200": {
            "description": "API keys da organização (sem hash — chave só aparece na criação)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKey"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "api-keys"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Cria uma API key (mp_live_…) com escopos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "posts:read",
                        "posts:write",
                        "channels:read",
                        "channels:write",
                        "media:write",
                        "analytics:read",
                        "webhooks:manage",
                        "mcp",
                        "mcp:read",
                        "mcp:write"
                      ]
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "name",
                  "scopes"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "a apiKey em claro aparece SOMENTE nesta resposta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": {
                      "type": "string"
                    },
                    "record": {
                      "$ref": "#/components/schemas/ApiKey"
                    }
                  },
                  "required": [
                    "apiKey",
                    "record"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "delete": {
        "tags": [
          "api-keys"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Revoga uma API key",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "chave revogada"
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/channels": {
      "get": {
        "tags": [
          "channels"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Lista os canais conectados (tokens nunca são expostos)",
        "responses": {
          "200": {
            "description": "canais",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Channel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/channels/{id}": {
      "delete": {
        "tags": [
          "channels"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Desconecta um canal (soft delete)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "canal desconectado"
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/posts": {
      "post": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Agenda um post (1 grupo → 1 publicação por canal)",
        "description": "Valida texto/mídia por canal e enfileira. `textByChannel` personaliza o texto do post principal por canal; `thread` cria réplicas encadeadas; `requireApproval` nasce DRAFT aguardando aprovação por link.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  },
                  "channelIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "minItems": 1,
                    "maxItems": 20
                  },
                  "publishAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timezone": {
                    "type": "string",
                    "default": "UTC"
                  },
                  "settingsByChannel": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "textByChannel": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 10000
                    }
                  },
                  "mediaIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "maxItems": 10
                  },
                  "thread": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "text": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 10000
                        },
                        "mediaIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "maxItems": 10
                        },
                        "delaySec": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 600
                        }
                      },
                      "required": [
                        "text"
                      ]
                    },
                    "maxItems": 24
                  },
                  "requireApproval": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "text",
                  "channelIds",
                  "publishAt"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "grupo agendado (ou DRAFT se requireApproval)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroup"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/posts/{groupId}": {
      "get": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Detalhe de um grupo de post",
        "description": "Além dos estados, expõe o conteúdo completo (texto base, texto/settings por canal e réplicas de thread) — é o que o composer usa p/ duplicar um post.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "grupo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroupDetail"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "posts"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Edita texto, horário e/ou settings por canal (re-agenda com nova versão de job)",
        "description": "Editar `text` sobrescreve o content de TODAS as publicações (overrides resetam). `settingsByChannel` (chave = channelId) faz merge nos settings de publicação do canal e é validado por provider.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "groupId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 10000
                  },
                  "publishAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "settingsByChannel": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "grupo re-agendado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostGroup"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "não encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/publications": {
      "get": {
        "tags": [
          "publications"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Feed flat de publicações (calendário/kanban)",
        "description": "Uma linha por publicação (o cliente agrupa por groupId). Cada item embute grupo (state/origin/awaitingApproval) e canal. Paginação keyset por (publishAt, id).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "início do período (ISO 8601)"
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "fim do período (ISO 8601)"
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "csv de estados de publicação",
              "example": "SCHEDULED,PUBLISHED"
            },
            "required": false,
            "name": "state",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "csv de ids de canal"
            },
            "required": false,
            "name": "channelId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "cursor keyset da página anterior"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "1–200 (default 100)",
              "example": "100"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "página do feed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationFeed"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Stream SSE de eventos em tempo real",
        "description": "Eventos nomeados: post.scheduled, post.published, post.failed, channel.refresh_required, notification.created (+ hello no handshake e ping a cada 25s). Sem Redis o stream fica só com keepalive — a UI cai no polling.",
        "responses": {
          "200": {
            "description": "stream text/event-stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}